Refactor out Specifier and RootSpecifier
