+++ src/schedule.rs
@@ -9,6 +9,7 @@ use std::str::{self, FromStr};
 
 use crate::time_unit::*;
 use crate::ordinal::*;
+use crate::specifier::*;
 
 #[derive(Clone, Debug)]
 pub struct Schedule {
@@ -597,34 +598,6 @@ where
             self.is_done = true;
             None
         }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub enum Specifier {
-    All,
-    Point(Ordinal),
-    Range(Ordinal, Ordinal),
-    NamedRange(String, String),
-}
-
-// Separating out a root specifier allows for a higher tiered specifier, allowing us to achieve
-// periods with base values that are more advanced than an ordinal:
-// - all: '*/2'
-// - range: '10-2/2'
-// - named range: 'Mon-Thurs/2'
-//
-// Without this separation we would end up with invalid combinations such as 'Mon/2'
-#[derive(Debug, PartialEq)]
-pub enum RootSpecifier {
-    Specifier(Specifier),
-    Period(Specifier, u32),
-    NamedPoint(String),
-}
-
-impl From<Specifier> for RootSpecifier {
-    fn from(specifier: Specifier) -> Self {
-        Self::Specifier(specifier)
     }
 }
